fix: prevent logging connection issue
With `start_with_request` set to `yes`, The log file is full of connection issue messages added while the debugger isn't turned on (which is the case most of the time). Changing the setting to `trigger` requires installing the xDebug browser extension but keeps our log file clean at the same time.